// src/interfaces/IJob.ts
export interface IJob {
  id: string
  code: string
  title: string
  description: string
  type: string
  positions: number
  experience: string
  salary: string
  industry: string
  location: any
  department: any
  division: any
  function: any
  postedDate: string
  closingDate: string
  hostedUrl: string
  applyUrl: string
}